+++ CHANGELOG.md
@@ -8,7 +8,25 @@ This changelog is following the recommended format by [keepachangelog](https://k
 
 ### Security
 
+
 ### Added
+
+
+### Changed
+
+
+### Fixed
+
+
+### Removed
+
+
+## [1.3.31] - 2026-09-25
+
+### Added
+
+- Add support for cloud rules by [@animesh-tarodia-sp](https://github.com/animesh-tarodia-sp) (cf. [#163](https://github.com/yannick-beot-sp/vscode-sailpoint-identitynow/pull/163))
+- Export a cloud rule or a `.java`/`.bsh` file as XML
 
 ### Changed
 
@@ -20,7 +38,6 @@ This changelog is following the recommended format by [keepachangelog](https://k
 - Fix issue with role dimension pagination (cf. [#164](https://github.com/yannick-beot-sp/vscode-sailpoint-identitynow/issues/164))
 - Issue when Importing Dimensions with bad entitlements (cf. [#160](https://github.com/yannick-beot-sp/vscode-sailpoint-identitynow/issues/160)). Fix issue with CacheService.
 
-### Removed
 
 ## [1.3.30] - 2026-07-06